implementing better oop design
